requirements.txt
flask==3.0.0
openai>=1.30.0
google-genai>=1.0.0
python-dotenv==1.0.0
httpx>=0.27.0
firebase-admin>=6.0.0
gunicorn>=21.0.0
